Changed Functors/Monads/Reader folder topology
