+++ src/Functors/Monads/Reader/index.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * reader monad functions
+ *
+ * @package bingo-functional
+ * @author Lochemem Bruno Michael
+ * @license Apache-2.0
+ */
+
+require_once __DIR__ . '/MapReader.php';
+require_once __DIR__ . '/Reader.php';
+require_once __DIR__ . '/RunReader.php';
+require_once __DIR__ . '/WithReader.php';
